+++ yuyo/reactions.py
@@ -76,15 +76,6 @@ class AbstractReactionHandler(abc.ABC):
     def has_expired(self) -> bool:
         """Whether this handler has ended."""
 
-    @property
-    @abc.abstractmethod
-    def last_triggered(self) -> datetime.datetime:
-        """When this handler was last triggered.
-
-        !!! note
-            If it hasn't ever been triggered then this will be when it was created.
-        """
-
     @abc.abstractmethod
     async def close(self) -> None:
         """Close this handler."""
@@ -126,7 +117,7 @@ class ReactionHandler(AbstractReactionHandler):
         self,
         *,
         authors: collections.Iterable[hikari.SnowflakeishOr[hikari.User]] = (),
-        timeout: datetime.timedelta = datetime.timedelta(seconds=30),
+        timeout: typing.Optional[datetime.timedelta] = datetime.timedelta(seconds=30),
     ) -> None:
         """Initialise a reaction handler.
 
@@ -159,17 +150,10 @@ class ReactionHandler(AbstractReactionHandler):
     @property
     def has_expired(self) -> bool:
         # <<inherited docstring from AbstractReactionHandler>>.
-        return self._timeout < datetime.datetime.now(tz=datetime.timezone.utc) - self._last_triggered
-
-    @property
-    def last_triggered(self) -> datetime.datetime:
-        # <<inherited docstring from AbstractReactionHandler>>.
-        return self._last_triggered
-
-    @property
-    def timeout(self) -> datetime.timedelta:
-        """How long this handler will wait since the last event before timing out."""
-        return self._timeout
+        return (
+            self._timeout is not None
+            and self._timeout < datetime.datetime.now(tz=datetime.timezone.utc) - self._last_triggered
+        )
 
     async def open(self, message: hikari.Message, /) -> None:
         self._message = message
@@ -294,7 +278,7 @@ class ReactionPaginator(ReactionHandler):
             pagination.STOP_SQUARE,
             pagination.RIGHT_TRIANGLE,
         ),
-        timeout: datetime.timedelta = datetime.timedelta(seconds=30),
+        timeout: typing.Optional[datetime.timedelta] = datetime.timedelta(seconds=30),
     ) -> None:
         """Initialise a reaction paginator.